Added the sequences given by Ru, and created a series of tests to use among other things.
